requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
numpy==1.26.4
opencv-python==4.9.0.80
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
